requirements.txt
playwright==1.41.0
pytest==8.0.0
pytest-playwright==0.4.4
pytest-asyncio==0.23.3
python-dotenv==1.0.0
boto3==1.34.34
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.0
openpyxl==3.1.2
pillow==10.2.0
